structural refactoring of io.sac.pysac into io.sac
