+++ obspy/io/sac/arrayio.py
