+++ obspy/io/sac/util.py
